+++ python/cudf/cudf/core/dataframe.py
@@ -3525,11 +3525,6 @@ class DataFrame(object):
     #
     # Stats
     #
-    def _as_common_dtype(self):
-        common_dtype = np.find_common_type(self.dtypes, [])
-        coerced = self._apply_support_method("astype", dtype=common_dtype)
-        return coerced
-
     def count(self, **kwargs):
         return self._apply_support_method("count", **kwargs)
 
@@ -3619,76 +3614,6 @@ class DataFrame(object):
             level=level,
             numeric_only=numeric_only,
         )
-
-    def cov(self, min_periods=None):
-        """Calculates the pairwise covariance of columns.
-        """
-        if not utils.IS_CUPY_AVAILABLE:
-            msg = (
-                "DataFrame.cov currently requires CuPy. "
-                "Please install CuPy to use this operation."
-            )
-            raise ImportError(msg)
-
-        import cupy as cp
-
-        assert min_periods in (None, 1)
-
-        if self.empty:
-            return self
-
-        if len(self) <= 1:
-            return np.nan
-
-        if any([col.has_null_mask for col in self._columns]):
-            msg = "Dataframe.cov does not currently support null values."
-            raise ValueError(msg)
-
-        filtered = self.select_dtypes(include=[np.number, np.bool])
-        coerced = filtered._as_common_dtype()
-        arr = cp.asarray(coerced.as_gpu_matrix())
-
-        result = cp.cov(arr, rowvar=False)
-        result_df = DataFrame.from_gpu_matrix(
-            cp.asfortranarray(result)
-        ).set_index(filtered.columns)
-        result_df.columns = filtered.columns
-        return result_df
-
-    def corr(self, method="pearson", min_periods=1):
-        """Calculates the pairwise covariance of columns.
-        """
-        if not utils.IS_CUPY_AVAILABLE:
-            msg = (
-                "DataFrame.corr currently requires CuPy. "
-                "Please install CuPy to use this operation."
-            )
-            raise ImportError(msg)
-
-        import cupy as cp
-
-        assert method in ("pearson",) and min_periods in (None, 1)
-
-        if self.empty:
-            return self
-
-        if len(self) <= 1:
-            return np.nan
-
-        if any([col.has_null_mask for col in self._columns]):
-            msg = "Dataframe.corr does not currently support null values."
-            raise ValueError(msg)
-
-        filtered = self.select_dtypes(include=[np.number, np.bool])
-        coerced = filtered._as_common_dtype()
-        arr = cp.asarray(coerced.as_gpu_matrix())
-
-        result = cp.corrcoef(arr, rowvar=False)
-        result_df = DataFrame.from_gpu_matrix(
-            cp.asfortranarray(result)
-        ).set_index(filtered.columns)
-        result_df.columns = filtered.columns
-        return result_df
 
     def all(self, bool_only=None, **kwargs):
         if bool_only: