+++ CHANGELOG.md
@@ -28,7 +28,7 @@
 - PR #2836 Add nvstrings.code_points method
 - PR #2844 Add Series/DataFrame notnull
 - PR #2858 Add GTest type list utilities
-- PR #2719 Series and DataFrame covariance and Pearson correlation
+- PR #2719 Series covariance and Pearson correlation
 
 ## Improvements